Corrected micros casting

+++ JsonTalkie/examples/ESP32_Broadcast_SPI_ESP_Master/S_Broadcast_SPI_2xESP_Master.hpp
@@ -66,7 +66,7 @@ protected:
 
 		static uint8_t stacked_transmissions = 0;
 
-		if (_in_time_slot == true && micros() - _sent_time_us > send_time_slot_us) {
+		if (_in_time_slot == true && (uint16_t)micros() - _sent_time_us > send_time_slot_us) {
 			_in_time_slot = false;
 		}
 
@@ -75,7 +75,7 @@ protected:
 		// Too many SPI sends to the Slaves asking if there is something to send will overload them, so, a timeout is needed
 		static uint16_t last_beacon_time = (uint16_t)micros();
 
-		if (micros() - last_beacon_time > 100) {
+		if ((uint16_t)micros() - last_beacon_time > 100) {
 			last_beacon_time = (uint16_t)micros();	// Avoid calling the beacon right away
 
 			if (_initiated) {
@@ -158,7 +158,7 @@ protected:
 				broadcastPayload(_spi_cs_pins, _ss_pins_count, (uint8_t)len);
 
 				_in_time_slot = true;
-				_sent_time_us = micros();
+				_sent_time_us = (uint16_t)micros();
 
 			} else {
 				return false;
